+++ src/main/java/com/planbase/pdf/layoutmanager/contents/Table.kt
@@ -21,7 +21,7 @@
 package com.planbase.pdf.layoutmanager.contents
 
 import com.planbase.pdf.layoutmanager.attributes.CellStyle
-import com.planbase.pdf.layoutmanager.attributes.DimAndPages
+import com.planbase.pdf.layoutmanager.attributes.DimAndPageNums
 import com.planbase.pdf.layoutmanager.attributes.TextStyle
 import com.planbase.pdf.layoutmanager.lineWrapping.LineWrappable
 import com.planbase.pdf.layoutmanager.lineWrapping.LineWrapped
@@ -102,17 +102,17 @@ class Table(val cellWidths:MutableList<Float> = mutableListOf(),
          * Renders item and all child-items with given width and returns the x-y pair of the
          * lower-right-hand corner of the last line (e.g. of text).
          */
-        override fun render(lp: RenderTarget, topLeft: Coord, reallyRender: Boolean): DimAndPages {
+        override fun render(lp: RenderTarget, topLeft: Coord, reallyRender: Boolean): DimAndPageNums {
             var y = topLeft.y
             var maxWidth = 0f
-            var pageNums:IntRange = DimAndPages.INVALID_PAGE_RANGE
+            var pageNums:IntRange = DimAndPageNums.INVALID_PAGE_RANGE
             for (part in parts) {
-                val dimAndPages: DimAndPages = part.render(lp, topLeft.y(y), reallyRender)
-                maxWidth = max(maxWidth, dimAndPages.dim.width)
-                y -= dimAndPages.dim.height
-                pageNums = dimAndPages.maxExtents(pageNums)
+                val dimAndPageNums: DimAndPageNums = part.render(lp, topLeft.y(y), reallyRender)
+                maxWidth = max(maxWidth, dimAndPageNums.dim.width)
+                y -= dimAndPageNums.dim.height
+                pageNums = dimAndPageNums.maxExtents(pageNums)
             }
-            return DimAndPages(Dim(maxWidth, topLeft.y - y), pageNums)
+            return DimAndPageNums(Dim(maxWidth, topLeft.y - y), pageNums)
         }
 
         override fun toString(): String = "WrappedTable($parts)"