+++ src/main/java/com/planbase/pdf/layoutmanager/contents/Text.kt
@@ -20,7 +20,7 @@
 
 package com.planbase.pdf.layoutmanager.contents
 
-import com.planbase.pdf.layoutmanager.attributes.DimAndPages
+import com.planbase.pdf.layoutmanager.attributes.DimAndPageNums
 import com.planbase.pdf.layoutmanager.attributes.TextStyle
 import com.planbase.pdf.layoutmanager.lineWrapping.ConTerm
 import com.planbase.pdf.layoutmanager.lineWrapping.ConTermNone
@@ -74,7 +74,7 @@ data class Text(val textStyle: TextStyle,
         }
 
         // Text rendering calculation spot 1/3
-        override fun render(lp: RenderTarget, topLeft: Coord, reallyRender: Boolean): DimAndPages =
+        override fun render(lp: RenderTarget, topLeft: Coord, reallyRender: Boolean): DimAndPageNums =
                 lp.drawStyledText(topLeft.minusY(ascent), string, textStyle, reallyRender)
                         .dimAndPagesFromWidth(dim)